requirements.txt
mysql-connector-python==8.0.31
pandas==1.5.2
pytelegrambotapi==4.9.0
python-dotenv==0.21.0
transformers==4.25.1
